+++ src/main/java/com/unascribed/yttr/mixin/accessor/client/AccessorClientPlayerInteractionManager.java
@@ -6,6 +6,7 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.network.ClientPlayerInteractionManager;
+import net.minecraft.util.math.BlockPos;
 
 @Environment(EnvType.CLIENT)
 @Mixin(ClientPlayerInteractionManager.class)
@@ -13,5 +14,10 @@ public interface AccessorClientPlayerInteractionManager {
 
 	@Accessor("blockBreakingCooldown")
 	void yttr$setBlockBreakingCooldown(int i);
+	
+	@Accessor("currentBreakingPos")
+	BlockPos yttr$getCurrentBreakingPos();
+	@Accessor("currentBreakingProgress")
+	float yttr$getCurrentBreakingProgress();
 	
 }