+++ src/main/java/com/unascribed/yttr/mixin/soul/client/MixinInGameHud.java
@@ -1,0 +1,156 @@
+package com.unascribed.yttr.mixin.soul.client;
+
+import java.util.concurrent.ThreadLocalRandom;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Coerce;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
+
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.BufferRenderer;
+import com.mojang.blaze3d.vertex.Tessellator;
+import com.mojang.blaze3d.vertex.VertexFormat.DrawMode;
+import com.mojang.blaze3d.vertex.VertexFormats;
+import com.unascribed.yttr.init.YBlocks;
+import com.unascribed.yttr.mixin.accessor.client.AccessorClientPlayerInteractionManager;
+import com.unascribed.yttr.mixin.accessor.client.AccessorHeartType;
+
+import static org.lwjgl.opengl.GL11.*;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.gui.DrawableHelper;
+import net.minecraft.client.gui.hud.InGameHud;
+import net.minecraft.client.render.GameRenderer;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.math.MathHelper;
+import net.minecraft.util.math.Vec3f;
+
+@Mixin(InGameHud.class)
+@Environment(EnvType.CLIENT)
+public class MixinInGameHud {
+
+	private float yttr$lastEvaporateProgress;
+	private float yttr$evaporateProgress;
+	private boolean yttr$evaporating;
+	private int yttr$wasEvaporating;
+	private boolean yttr$hide;
+	private float yttr$lastMaxHealth;
+	private int yttr$heartResolution;
+	private Vec3f[] yttr$hearticulates; // x, y, start
+	
+	@Inject(at=@At("HEAD"), method="tick")
+	private void yttr$tickEvaporate(boolean paused, CallbackInfo ci) {
+		var mc = MinecraftClient.getInstance();
+		yttr$lastEvaporateProgress = yttr$evaporateProgress;
+		if (mc.world != null && mc.interactionManager instanceof AccessorClientPlayerInteractionManager acc
+				&& mc.interactionManager.isBreakingBlock()
+				&& mc.world.getBlockState(acc.yttr$getCurrentBreakingPos()).isOf(YBlocks.POLISHED_SCORCHED_OBSIDIAN_HOLSTER)) {
+			yttr$evaporateProgress = acc.yttr$getCurrentBreakingProgress();
+			yttr$wasEvaporating = yttr$evaporateProgress > 0.95 ? 30 : 0;
+			yttr$lastMaxHealth = mc.player.getMaxHealth();
+		} else {
+			yttr$evaporateProgress = 0;
+			yttr$heartResolution = 0;
+			yttr$hearticulates = null;
+		}
+		if (yttr$evaporateProgress > 0) {
+			if (yttr$hearticulates == null) {
+				mc.getTextureManager().bindTexture(DrawableHelper.GUI_ICONS_TEXTURE);
+				yttr$heartResolution = (glGetTexLevelParameteri(GL_TEXTURE_2D, 0, GL_TEXTURE_WIDTH)*9)/256;
+				yttr$hearticulates = new Vec3f[yttr$heartResolution*yttr$heartResolution];
+				var r = ThreadLocalRandom.current();
+				for (int i = 0; i < yttr$hearticulates.length; i++) {
+					float y = ((i/yttr$heartResolution)+1)/(float)yttr$heartResolution;
+					yttr$hearticulates[i] = new Vec3f((float)r.nextGaussian(0, 4), -r.nextFloat(12), r.nextFloat()*r.nextFloat(y));
+				}
+			}
+		}
+		yttr$wasEvaporating--;
+	}
+	
+	@Inject(at=@At(value="FIELD", target="net/minecraft/client/gui/hud/InGameHud$HeartType.CONTAINER:Lnet/minecraft/client/gui/hud/InGameHud$HeartType;"),
+			method="renderHealthBar", locals=LocalCapture.CAPTURE_FAILHARD)
+	private void yttr$contextualizeEvaporatingHeart(MatrixStack matrices, PlayerEntity player, int x, int y, int lines, int regeneratingHeartIndex, float maxHealth, int lastHealth, int health, int absorption, boolean blinking,
+			CallbackInfo ci, @Coerce Object unused0, int unused1, int maxHearts, int unused2, int unused3, int i) {
+		yttr$evaporating = (yttr$hearticulates != null && i == (maxHearts-1));
+		yttr$hide = false;
+		if (!yttr$evaporating && yttr$wasEvaporating > 0) {
+			float cutoff = player.getMaxHealth();
+			if (player.getMaxHealth() == yttr$lastMaxHealth) {
+				cutoff -= 2;
+			}
+			if ((i*2) >= cutoff) {
+				yttr$hide = true;
+			}
+		}
+	}
+	
+	@Inject(at=@At("HEAD"), method="drawHeart", cancellable=true)
+	private void yttr$drawEvaporatingHeart(MatrixStack matrices, @Coerce AccessorHeartType type, int x, int y, int v, boolean blinking, boolean halfHeart, CallbackInfo ci) {
+		if (yttr$hide) {
+			ci.cancel();
+			return;
+		}
+		if (yttr$evaporating) {
+			var mc = MinecraftClient.getInstance();
+			var bldr = Tessellator.getInstance().getBufferBuilder();
+			var mat = matrices.peek().getModel();
+			float f = 9f/yttr$heartResolution;
+			int u = type.yttr$getU(halfHeart, blinking);
+			float prog = MathHelper.lerp(mc.getTickDelta(), yttr$lastEvaporateProgress, yttr$evaporateProgress);
+			RenderSystem.enableBlend();
+			RenderSystem.defaultBlendFunc();
+			RenderSystem.setShader(GameRenderer::getPositionTexColorShader);
+			RenderSystem.setShaderColor(1, 1, 1, 1);
+			bldr.begin(DrawMode.QUADS, VertexFormats.POSITION_TEXTURE_COLOR);
+			for (int xi = 0; xi < yttr$heartResolution; xi++) {
+				for (int yi = 0; yi < yttr$heartResolution; yi++) {
+					var params = yttr$hearticulates[(yi*yttr$heartResolution)+xi];
+					float start = params.getZ();
+					float a;
+					if (prog > start) {
+						float span = 1-start;
+						if (span != 1) {
+							a = 1-((prog-start)/span);
+						} else {
+							a = prog;
+						}
+					} else {
+						a = 1;
+					}
+					if (a <= 0) continue;
+					float ai = 1-a;
+					float xo = params.getX()*ai;
+					float yo = params.getY()*ai;
+					
+					float xif = xi*f;
+					float yif = yi*f;
+					
+					float x0 = x+(xif)+xo;
+					float x1 = x0+f;
+					float y0 = y+(yif)+yo;
+					float y1 = y0+f;
+					
+					float u0 = (u + (xif)) / 256f;
+					float u1 = (u + (xif+f)) / 256f;
+					float v0 = (v + (yif)) / 256f;
+					float v1 = (v + (yif+f)) / 256f;
+
+					bldr.vertex(mat, x0, y1, 0).uv(u0, v1).color(1, 1, 1, a).next();
+					bldr.vertex(mat, x1, y1, 0).uv(u1, v1).color(1, 1, 1, a).next();
+					bldr.vertex(mat, x1, y0, 0).uv(u1, v0).color(1, 1, 1, a).next();
+					bldr.vertex(mat, x0, y0, 0).uv(u0, v0).color(1, 1, 1, a).next();
+				}
+			}
+			BufferRenderer.drawWithShader(bldr.end());
+			ci.cancel();
+		}
+	}
+
+}